Revert "changed afm type to Long"
This reverts commit 8daed8ebfced7a599b6cc38d6d0291c818ec543d.

+++ src/main/java/gr/hua/distributed/it21774/controller/UserContractController.java
@@ -111,9 +111,9 @@ public class UserContractController {
         Contract contract = new Contract(text, dateCreated, "", "");
 
         // Associate contract to users
-        Set<Long> afm = createContractRequest.getAfm();
+        Set<Integer> afm = createContractRequest.getAfm();
 
-        for (Long tempAfm : afm) {
+        for (Integer tempAfm : afm) {
 
             AppUser appUser2;
 
@@ -128,7 +128,7 @@ public class UserContractController {
             }
         }
 
-        for (Long tempAfm : afm) {
+        for (Integer tempAfm : afm) {
             AppUser appUser3 = appUserRepository.findByAfm(tempAfm).get();
             contract.addAppUsers(appUser3);
         }
